Fixed up tests and renamed `fabric-value-modern.ts` -> `type-check.ts`.
On the latter front, there was only the one function left after
splitting out `native-conversion.ts`.
